Fix golint warnings

+++ relay/pkg/ethash/sizes.go
@@ -11,7 +11,7 @@ func cacheSize(epoch uint64) uint64 {
 	}
 
 	// No known cache size, calculate manually (sanity branch only)
-	size := cacheInitBytes + cacheGrowthBytes*uint64(epoch) - hashBytes
+	size := cacheInitBytes + cacheGrowthBytes*epoch - hashBytes
 	for !new(big.Int).SetUint64(size / hashBytes).ProbablyPrime(1) { // Always accurate for n < 2^64
 		size -= 2 * hashBytes
 	}
@@ -27,7 +27,7 @@ func datasetSize(epoch uint64) uint64 {
 		return datasetSizes[epoch]
 	}
 
-	size := datasetInitBytes + datasetGrowthBytes*uint64(epoch) - mixBytes
+	size := datasetInitBytes + datasetGrowthBytes*epoch - mixBytes
 	for !new(big.Int).SetUint64(size / mixBytes).ProbablyPrime(1) { // Always accurate for n < 2^64
 		size -= 2 * mixBytes
 	}
